BJ-1004: Refactor to custom metrics starter
