+++ pom.xml
@@ -10,13 +10,13 @@
         <version>1.0.5</version>
     </parent>
 
-    <artifactId>spring-boot-starter-metrics-statsd</artifactId>
-    <version>1.1.3</version>
+    <artifactId>custom-metrics-spring-boot-starter</artifactId>
+    <version>1.1.4</version>
     <packaging>jar</packaging>
 
-    <name>Spring boot starter metrics statsd</name>
-    <description>Starter for statsd metrics</description>
-    <url>https://github.com/rbkmoney/spring-boot-starter-metrics-statsd.git</url>
+    <name>Custom metrics Spring Boot starter</name>
+    <description>Starter for StatsD and Prometheus custom metrics</description>
+    <url>https://github.com/rbkmoney/custom-metrics-spring-boot-starter.git</url>
 
     <developers>
         <developer>
@@ -28,14 +28,15 @@
     </developers>
 
     <scm>
-        <connection>scm:git:git://github.com/rbkmoney/spring-boot-starter-metrics-statsd.git</connection>
-        <developerConnection>scm:git:ssh://github.com/rbkmoney/spring-boot-starter-metrics-statsd.git</developerConnection>
-        <url>https://github.com/rbkmoney/spring-boot-starter-metrics-statsd/tree/master</url>
+        <connection>scm:git:git://github.com/rbkmoney/custom-metrics-spring-boot-starter.git</connection>
+        <developerConnection>scm:git:ssh://github.com/rbkmoney/custom-metrics-spring-boot-starter.git</developerConnection>
+        <url>https://github.com/rbkmoney/custom-metrics-spring-boot-starter/tree/master</url>
     </scm>
 
     <properties>
         <spring.version>2.3.4.RELEASE</spring.version>
         <java.version>8</java.version>
+        <micrometer.version>1.5.6</micrometer.version>
     </properties>
 
     <dependencies>
@@ -56,7 +57,6 @@
             <version>${spring.version}</version>
             <scope>provided</scope>
         </dependency>
-
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-actuator</artifactId>
@@ -65,7 +65,17 @@
         <dependency>
             <groupId>io.micrometer</groupId>
             <artifactId>micrometer-registry-statsd</artifactId>
-            <version>1.3.1</version>
+            <version>${micrometer.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.micrometer</groupId>
+            <artifactId>micrometer-core</artifactId>
+            <version>${micrometer.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.micrometer</groupId>
+            <artifactId>micrometer-registry-prometheus</artifactId>
+            <version>${micrometer.version}</version>
         </dependency>
     </dependencies>
